+++ server/src/main/java/org/realityforge/replicant/server/transport/ReplicantSession.java
@@ -6,7 +6,6 @@ import java.util.Map;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import org.realityforge.replicant.server.ChannelDescriptor;
-import org.realityforge.ssf.SimpleSessionInfo;
 
 /**
  * Base class for sessions within replicant.